SY-2937: Reorg part 1
