+++ integration/test/console/console.py
@@ -21,8 +21,7 @@ class Console(TestCase):
     """
     Console TestCase implementation using Playwright
 
-    SY-2965: Break out child classes to be independent console panel handlers
-    as opposed to individual console tests
+    SY-2965: Break this out to Console & ConsoleCase
     """
 
     browser: Browser
