+++ integration/test/console/plot/plot.py
