+++ integration/test/console/schematic/__init__.py
@@ -1,0 +1,15 @@
+#  Copyright 2025 Synnax Labs, Inc.
+#
+#  Use of this software is governed by the Business Source License included in the file
+#  licenses/BSL.txt.
+#
+#  As of the Change Date specified in that file, in accordance with the Business Source
+#  License, use of this software will be governed by the Apache License, Version 2.0,
+#  included in the file licenses/APL.txt.
+
+from .schematic import Schematic
+from .schematic_symbol import SchematicSymbol
+from .setpoint import Setpoint
+from .value import Value
+
+__all__ = ["Schematic", "SchematicSymbol", "Value", "Setpoint"]
